requirements.txt
python-dotenv~=0.14.0
discord~=1.0.1
oauth2client~=4.1.3
google-api-python-client~=1.10.0